fix size_t overflow of InsertionSort

+++ ch02/include/ch02/insertion_sort.h
@@ -1,13 +1,11 @@
 #pragma once
 
-#include <cstddef>
-
 namespace ch02 {
 template <typename Container>
 void InsertionSortNonDecrease(Container &seq) {
-  for (size_t j = 0; j < seq.size(); ++j) {
+  for (int j = 0; j < (int)seq.size(); ++j) {
     auto key = seq[j];
-    size_t i = j - 1;
+    int i = j - 1;
     while (i >= 0 && seq[i] > key) {
       seq[i + 1] = seq[i];
       i = i - 1;
@@ -18,9 +16,9 @@ void InsertionSortNonDecrease(Container &seq) {
 
 template <typename Container>
 void InsertionSortNonIncrease(Container &seq) {
-  for (size_t j = 0; j < seq.size(); ++j) {
+  for (int j = 0; j < (int)seq.size(); ++j) {
     auto key = seq[j];
-    size_t i = j - 1;
+    int i = j - 1;
     while (i >= 0 && seq[i] < key) {
       seq[i + 1] = seq[i];
       i = i - 1;
